move OnPlatformExtensions to Controls library Rename OnPlatformEx to OnPlatform because it's namespaced.
