+++ src/Consolonia.Controls/Assembly.cs
@@ -4,5 +4,6 @@ using Avalonia.Metadata;
 [assembly: CLSCompliant(false)] //todo: should we make it compliant?
 
 [assembly: XmlnsDefinition("https://github.com/jinek/consolonia", "Consolonia.Controls")]
-[assembly: XmlnsDefinition("https://github.com/jinek/consolonia", "Consolonia.Controls.Dialog")]
 [assembly: XmlnsDefinition("https://github.com/jinek/consolonia", "Consolonia.Controls.Brushes")]
+[assembly: XmlnsDefinition("https://github.com/jinek/consolonia", "Consolonia.Controls.Dialog")]
+[assembly: XmlnsDefinition("https://github.com/jinek/consolonia", "Consolonia.Controls.Markup")]
