+++ src/Consolonia.Controls/Consolonia.Controls.csproj
@@ -7,4 +7,8 @@
         <PackageReference Include="CommunityToolkit.Mvvm" Version="8.3.2" />
     </ItemGroup>
 
+    <ItemGroup>
+      <Folder Include="Markup\" />
+    </ItemGroup>
+
 </Project>
